+++ tests/listedit.rs
@@ -3,6 +3,9 @@
 #![feature(zero_one)]
 // #![plugin(quickcheck_macros)]
 
+#[macro_use] extern crate log;
+
+
 //
 // cargo test listedit::experiments -- --nocapture
 //
@@ -17,10 +20,61 @@ use std::num::Zero;
 use adapton::adapton_sigs::* ;
 use adapton::collection_traits::*;
 use adapton::collection_edit::*;
+use adapton::collection::*;
 use adapton::engine;
 use adapton::naive;
+use std::fmt::Debug;
+use std::hash::Hash;
 
 type Edits = Vec<CursorEdit<u32, Dir2>>;
+
+fn has_consecutive_names<A:Adapton,X,L:ListT<A,X>> (st:&mut A, list:L::List) -> bool {
+    L::elim(st, list,
+            |st| false,
+            |st,x,xs| has_consecutive_names::<A,X,L> (st, xs),
+            |st,n,xs|
+            L::elim(st, xs,
+                    |st| false,
+                    |st,y,ys| has_consecutive_names::<A,X,L> (st, ys),
+                    |st,m,ys| true))
+}
+
+pub struct Experiment ;
+impl<A:Adapton,X:Zero+Hash+Debug+PartialEq+Eq+Clone+PartialOrd> ExperimentT<A,X,Vec<X>>
+    for Experiment
+{
+    type ListEdit = ListZipper<A,X,List<A,X>> ;
+    fn run (st:&mut A, edits:Vec<CursorEdit<X,Dir2>>, view:ListReduce) -> Vec<(Vec<X>,Cnt)> {
+        debug!("run");
+        let mut outs : Vec<(Vec<X>,Cnt)> = Vec::new();
+        let mut z : ListZipper<A,X,List<A,X>> = Self::ListEdit::empty(st) ;
+        let mut loop_cnt = 0 as usize;
+        for edit in edits.into_iter() {
+            debug!("\n----------------------- Loop head; count={}", loop_cnt);
+            debug!("zipper: {:?}", z);
+            let consecutive_left  = has_consecutive_names::<A,X,List<A,X>>(st, z.left.clone());
+            let consecutive_right = has_consecutive_names::<A,X,List<A,X>>(st, z.right.clone());
+            debug!("zipper names: consecutive left: {}, consecutive right: {}",
+                     consecutive_left, consecutive_right);
+            //assert!(!consecutive_left);  // Todo-Later: This assertion generally fails for random interactions
+            //assert!(!consecutive_right); // Todo-Later: This assertion generally fails for random interactions
+            debug!("edit:   {:?}", edit);
+            let (out, cnt) = st.cnt(|st|{
+                let z_next = eval_edit::<A,X,Self::ListEdit>(st, edit, z.clone(), loop_cnt);
+                let tree = Self::ListEdit::get_tree::<Tree<A,X,u32>>(st, z_next.clone(), Dir2::Left);
+                debug!("tree:   {:?}", tree);
+                let nm = st.name_of_string("eval_reduce".to_string());
+                let out = st.ns(nm, |st|eval_reduce::<A,X,Tree<A,X,u32>>(st, tree, &view) );
+                z = z_next;
+                loop_cnt = loop_cnt + 1;
+                out
+            }) ;
+            debug!("out:    {:?}", out);
+            debug!("cnt:    {:?}", cnt);
+            outs.push((out,cnt));
+        } outs
+    }
+}
 
 fn compare_naive_and_cached(edits: &Edits) -> bool {
     let reduction = ListReduce::Max;
